# save-backend/backend-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SAVE Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5800",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/projects/": {
      "get": {
        "description": "Get all projects, available for current user.",
        "operationId": "getProjects",
        "parameters": [
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Projects successfully fetched."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get all available projects.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/all": {
      "get": {
        "description": "Get all projects, including deleted and private. Only accessible for super admins",
        "operationId": "getProjects_1",
        "parameters": [
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Projects successfully fetched."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get all projects.",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/get/organization-name": {
      "get": {
        "description": "Get project by name and organization name.",
        "operationId": "getProjectByNameAndOrganizationName",
        "parameters": [
          {
            "description": "name of a project",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successfully fetched project by name and organization name."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Not enough permission for accessing given project."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Could not find project with such name and organization name."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get project by name and organization name.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/get/projects-by-organization": {
      "get": {
        "description": "Get all projects by organization name.",
        "operationId": "getProjectsByOrganizationName",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Successfully fetched projects by organization name."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get all projects by organization name.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/not-deleted": {
      "post": {
        "description": "Get non-deleted projects, available for current user.",
        "operationId": "getNotDeletedProjectsWithFilters",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectFilters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Successfully fetched non-deleted projects."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get non-deleted projects.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/save": {
      "post": {
        "description": "Create a new project.",
        "operationId": "saveProject",
        "parameters": [
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successfully created a new project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Could not find organization with such name."
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Either invalid data, or project with such name is already created."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Create a new project.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/update": {
      "post": {
        "description": "Update an existing project.",
        "operationId": "updateProject",
        "parameters": [
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successfully updated a project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Not enough permission for managing project settings."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Either project or organization could not be found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Update an existing project.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{organizationName}/{projectName}/delete": {
      "delete": {
        "description": "Delete a project.",
        "operationId": "deleteProject",
        "parameters": [
          {
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successfully deleted a project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Not enough permission for project deletion."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Either could not find such organization or such project in such organization."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Delete a project.",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/users/{userName}": {
      "get": {
        "operationId": "findByName",
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "users-details-controller"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "required": [
          "canCreateContests",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string"
          },
          "canCreateContests": {
            "type": "boolean"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "DELETED"
            ]
          }
        }
      },
...
      "Project": {
        "required": [
          "contestRating",
          "name",
          "numberOfContainers",
          "organization",
          "public",
          "status"
        ],
        "type": "object",
        "properties": {
          "contestRating": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "numberOfContainers": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "public": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "DELETED"
            ]
          },
          "url": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProjectDto": {
        "required": [
          "description",
          "email",
          "isPublic",
          "name",
          "organizationName",
          "url"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ProjectFilters": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "UserInfo": {
        "required": [
          "isActive",
          "name",
          "organizations",
          "originalLogins",
          "projects"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "gitHub": {
            "type": "string"
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "NONE",
              "OWNER",
              "SUPER_ADMIN",
              "VIEWER"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "isActive": {
            "type": "boolean"
          },
          "linkedin": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "oldName": {
            "type": "string"
          },
          "organizations": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ADMIN",
                "NONE",
                "OWNER",
                "SUPER_ADMIN",
                "VIEWER"
              ]
            }
          },
          "originalLogins": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projects": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ADMIN",
                "NONE",
                "OWNER",
                "SUPER_ADMIN",
                "VIEWER"
              ]
            }
          },
          "source": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  }
}